fix(kueue): move the arming latch to kueue_preflight, under the file-size cap
`Server Guards` failed: the latch and its test pushed config.rs to 54131 bytes,
over the 51200-byte cap. Adding a `djinn:allow-oversize` marker would have been
the wrong fix — the latch is preflight state, not config schema, so it belongs
next to the check that sets it. config.rs keeps only the one-line consult and
drops back to 50312 bytes without any marker.

The latch test moves with it. It shares `config::tests::ENV_LOCK` (now
`pub(crate)`) rather than declaring its own mutex: it drives
`KubernetesConfig::from_env()`, so a second lock would serialize nothing against
the sibling env tests and both would race on `DJINN_KUEUE_ARMED`.

Also documents what the startup check deliberately does not cover: it runs once
at boot, so stripping the label from a live armed deployment is not detected
until restart. Re-checking on a timer could not unsuspend the Jobs it would
belatedly disapprove of, and would add a way for a transient apiserver blip to
disarm a healthy cluster mid-flight.

Mutation re-verified after the move: deleting the `kueue_disarmed_by_preflight()`
branch in `from_env` still fails `a_preflight_refusal_disarms_every_later_from_env`.
